x402/requirements.txt
flask>=2.0.0
requests>=2.25.0
pynacl>=1.5.0
mnemonic>=0.20
